Test/redux (#63)
* add test to chart and headline components

* split up all other components reducer test

* refactor test

* move global initial state

* add spread operator
